test: Move MockRegister to a shared file
